refactor: resolved issue #2
