+++ src/lib/prepareInput/prepareInput.ts
@@ -1,11 +1,7 @@
-import {
-  INITIAL_CITY_COINS_COUNT,
-  MAX_X_COORDINATE,
-  MAX_Y_COORDINATE,
-} from '../../config';
 import { City, Country } from '../../types';
 import { Input } from '../readInput/readInput';
-import { createMatrix } from '../../utils/createMatrix';
+import { createEUMatrix } from './createEUMatrix';
+import { createCountriesWithCities } from './createCountriesWithCities';
 
 export type EU = {
   matrix: City[][];
@@ -14,36 +10,9 @@ export type EU = {
 
 export const prepareInput = (input: Input): EU[] => {
   const EUs: EU[] = input.map(({ countries }) => {
-    const matrix = createMatrix<City>(
-      MAX_X_COORDINATE + 1,
-      MAX_Y_COORDINATE + 1,
-      (x, y) => ({
-        x,
-        y,
-        coins: {
-          count: countries.reduce((acc, country) => {
-            acc[country.name] = 0;
-            return acc;
-          }, {} as Record<string, number>),
-        },
-        coinsToReceive: {
-          count: {},
-        },
-        country: '',
-      }),
-    );
-    const countriesWithCities = countries.map(country => {
-      const cities: City[] = [];
-      for (let x = country.xl; x <= country.xh; x++) {
-        for (let y = country.yl; y <= country.yh; y++) {
-          const city = matrix[x][y];
-          city.coins.count[country.name] = INITIAL_CITY_COINS_COUNT;
-          city.country = country.name;
-          cities.push(city);
-        }
-      }
-      return { ...country, cities };
-    });
+    const matrix = createEUMatrix(countries);
+    const countriesWithCities = createCountriesWithCities(countries, matrix);
+
     return {
       matrix,
       countries: countriesWithCities,
