+++ src/lib/readInput/readInput.ts
@@ -1,6 +1,6 @@
 import { Country } from '../../types';
 import { getherUserInputLines } from './getherUserInputLines';
-import { parseInput } from './parseInput';
+import { parseInput } from './parseInput/parseInput';
 
 export type Input = {
   countries: Omit<Country, 'cities'>[];
